BUGFIXING clear cache folders only if dir exists
Wrapping the clearDirectory($dir) methods logic in IF statement to only let the RecursiveDirectoryIterator delete the cache directories if the given directory already exists. This will fix a bug occuring when trying to regenerate the cache by console command. if the themes html directory isn't existing an exception will be thrown.

+++ engine/Shopware/Components/CacheManager.php
@@ -399,24 +399,26 @@ class CacheManager
      */
     private function clearDirectory($dir)
     {
-        $iterator = new \RecursiveIteratorIterator(
-            new \RecursiveDirectoryIterator($dir, \RecursiveDirectoryIterator::SKIP_DOTS),
-            \RecursiveIteratorIterator::CHILD_FIRST
-        );
+        if (file_exists($dir)) {
+            $iterator = new \RecursiveIteratorIterator(
+                new \RecursiveDirectoryIterator($dir, \RecursiveDirectoryIterator::SKIP_DOTS),
+                \RecursiveIteratorIterator::CHILD_FIRST
+            );
 
-        /** @var \SplFileInfo $path */
-        foreach ($iterator as $path) {
-            if ($path->getFilename() === '.gitkeep') {
-                continue;
-            }
-
-            if ($path->isDir()) {
-                rmdir($path->__toString());
-            } else {
-                if (!$path->isFile()) {
+            /** @var \SplFileInfo $path */
+            foreach ($iterator as $path) {
+                if ($path->getFilename() === '.gitkeep') {
                     continue;
                 }
-                unlink($path->__toString());
+
+                if ($path->isDir()) {
+                    rmdir($path->__toString());
+                } else {
+                    if (!$path->isFile()) {
+                        continue;
+                    }
+                    unlink($path->__toString());
+                }
             }
         }
     }
